+++ test/test_kernel_sort.cc
@@ -53,10 +53,10 @@ TEST_CASE("Test sorting", "[kernel::sort]")
 TEST_CASE("Sorting benchmark", "[kernel::sort]")
 {
     metalchat::device gpu0("metalchat.metallib", 1);
-    metalchat::sort<bf16, 128> sort(gpu0);
+    metalchat::sort<float, 128> sort(gpu0);
 
-    auto input_cpu = rand<bf16>({1, 1, 128256});
-    auto input = shared_tensor(empty<bf16>({1, 1, 128256}, gpu0));
+    auto input_cpu = rand<float>({1, 1, 128256});
+    auto input = shared_tensor(empty<float>({1, 1, 128256}, gpu0));
     std::copy(input_cpu.begin(), input_cpu.end(), input.begin());
 
     std::cout << "--running--" << std::endl;
